requirements.txt
kafka-python==2.0.2
kafka-python-ng==2.2.2
logging==0.4.9.6
